🍞 重复刷新什么的我觉得不行

+++ lib/oneindex.php
@@ -25,17 +25,23 @@
 		}
 		/*中转refreshcache，原refreshcache写法有多层循环*/
         static function refresh_cache($path){
+		   if(!file_exists(dirname(__FILE__).'/refresh.lock')){
 			require dirname(__FILE__).'/../config/refreshfix.php';
 			date_default_timezone_set("Asia/Shanghai");
 			if(time()>=intval($rconfig['nextrefresh'])){
+			   file_put_contents(dirname(__FILE__).'/refresh.lock','Refreshing');
 			   $rconfig['nextrefresh']=time()+intval($rconfig['refreshinterval']);
 			   file_put_contents(dirname(__FILE__).'/../config/refreshfix.php','<?php $rconfig='.var_export($rconfig,true).';?>');/*储存一遍时间*/
 			   $rt=self::real_refresh_cache($path);
 			   file_put_contents(dirname(__FILE__).'/../lastupdate.txt',date('Y-m-d h:i:sa',time()));
+			   unlink(dirname(__FILE__).'/refresh.lock');
 			   return $rt;
 			}else{/*未到刷新时间*/
 				return 'timefailed';
 			}
+		   }else{/*刷新中*/
+		       return 'refreshing'; 
+		   }
 		}
 		// 真-刷新缓存
 		static function real_refresh_cache($path){
